EsStorage loads the docs fully (and just some fields)
Unfortunately the usage of POStores for doing projections is
unreliable since they only provide the information for that particular
operation and not the actual schema. As the actual field names cannot
be properly resolved, this leads to invalid queries to ES and thus
incorrect results.

The only way forward seem to be passing the schema manually to the
schema itself as there does not seem to be any other reliable way of
doing this.

relates to #209
relates to #117