Add Apache Spark integration
Provides dedicated RDDs (one for Scala and one for Java) for reading data from Elasticsearch.
Additionally one can write any custom RDD to Elasticsearch as long as its content can be mapped to JSON
(if that is not the case, one can easily transform the data or provide their own customer serializer).