Fix correct 'pinning' of tasks to target node
Due to the way the settings are used, the pinned node for a certain
task is being ignored and the list of discovered nodes is used
instead. This commit addresses that and separates the internal
properties used by the library to avoid overloading them (and thus
leading to errors).

Fix #256