[SPARK] Upgrade to Spark 1.3
Unfortunately Spark 1.3 broke backwards compatibility with 1.2
in the SQL package. While the core is the same and serialization works
across both versions, the user API is simply different and the user
needs to pick one over the other.

To keep supporting pre-Spark 1.3 users, the project provides two
binaries - one for Spark 1.0-1.2 and one for Spark 1.3. This applies
only to the SQL package as the core is the same but to avoid
confusion and keep things simple, the recommendation is for users to
pick one over the other.