Overhaul for Geo types
Add dedicated parsing and handling of Geo types and inferring of data
based on 'sampling' of data.
As Geo types are not properly described into their mappings (ES provides
only `geo_shape` and `geo_point` but there's no information about the
geo type used), ES-Hadoop now detects a geo field and will parse it in an
ad-hoc manner.
However for strongly-typed environments (such as Spark SQL), it will
'sample' the data, by asking for one document so the actual content will
be parsed in order to determine the format and use that for the inferred
data set.

relates #607