Handle filter wrapped in a query in RawQueryBuilder
For elasticsearch versions prior to 5.x the raw query builder should be able to parse a query that wraps a filter.
Since we use the RawQueryBuilder to parse root queries as well we need to differenciate the parsing of a query and the parsing of a filter.
For the former we need to remove the root "query" object if it exists and for the latter we need to keep it because it only means that the filter is wrapped in a query.
This change also adds javadoc for the query builders.