Restores the ability to prefer local nodes when creating a PartitionReader
In Hadoop and Spark it is possible to set the list of hostnames/ips that should be preferred when a task is launched.
This change restores the heuristic that prefers ES nodes that are local to the task when a partition reader is created.
This means that if an ES node is on the same machine than an Hadoop/Spark worker and this node contains the targeted index/shardId then the PartitionReader will request this node in priority.