[REST] Adding support for Unicode characters in index and type names.
The SimpleRequest object takes both path and query parts for the URL, but so much of the framework combines them together into the path for ease of use. SimpleRequest now senses this, and splits them apart so that they may be each encoded separately from each other (queries are often encoded at time of their creation, but paths are encoded just before the request is sent).
fixes #857