Create testing fixture to test using Elasticsearch spun up by Gradle instead of embedded.
Downloading 6.0 snapshot artifacts now
Adding entries to major version
parsing version 6.0
set project version to alpha 1
Really sloppy first pass of getting a test fixture that runs ES loaded. This is heavily based on the existing test cluster standup in Elasticsearch, but with most of the functionality torn out of it
Setup and tear down of ES node works.
Upgrading Gradle to 3.3
Add wait task to ensure cluster is started before sending requests.
Pull all gradle distro items instead of only the bin
Revert back to 5.0.0 dependencies to ensure that it works on the expected version of ES
Gut the EsEmbeddedServer until I can hook up the ES integration tests
Test runs for MR package.
Reorganize the build file
Throw out the commented code, and some old logic from embedded server and local es.
Extract the cluster create into it's own buildscript.
EmbeddedServer to read a ports file from the system properties and add it as the address for the node.
LocalES has been slimmed down to either log that there is no fixture node or to pull and configure the ports file for the fixture node through the embedded server class.