[SPARK] Adding alpha support for Spark Structured Streaming
DataSource now provides the ability to return an Elasticsearch based Sink
for Structured Streaming outputs. The sink implementation is based on the
same serialization logic from the Spark SQL packages, using a RowEncoder
to convert the internal row representation into something that the SQL
packages already know how to handle. The Sink is backed by an HDFS client
based compacting commit log, which stands to try and insulate the sink
from writing data multiple times. Included with this commit is
documentation for the new feature, as well as a suite of tests using an
in-house built testing harness for structured streaming.