Fix spark sql queries with metadata field (#930)
The PR solves `scala.MatchError` in Spark SQL queries with `_metadata` field inside.
When Spark SQL is reading a metadata field, the data source removes the metadata
field from the given fields and replaces it at the end. This is because the metadata
field is an abstract field that is provided by the Scroll Reader instead of by explicitly
asking for it. In removing the field, the order of fields in the projection returned from
the Scroll reader does not match the order of fields within the Spark execution plan,
thus a match error is thrown. This ordering of fields is now protected with this PR.