Fixing `Malformed scroll id` errors in the scroll reading code.
In some cases, especially prevalent in jobs with large numbers of index
partition slices, a slice can be empty on first read since no records
have hashed into it. In these cases the scroll reader returns an empty
scroll, with size of -1 and an empty scroll id. If es.scroll.limit is set
then the empty batch read from the scroll is ignored and replaced with
the limit value. At this point the reader will attempt to read the
scroll again even though it is empty, but this time with a blank id.
This commit makes it so that if a scroll request is empty on the first
read, it is marked as finished immediately. The scroll id is now also
always returned, even on an empty scroll so that the appropriate close
calls can be made.

Fixes #935