Added exceptions when encountering TTL and Timestamp mappings on ES 6+
InitializationUtils validates that the settings are not present when
running against ES 6+, and logs deprecation warnings if they are
present when running against ES 5.x and below. Added deprecated tags
to the TTL and Timestamp metadata enums (in MetadataExtractor AND in
spark rdd packages). When using the metadata tags with Spark RDD's
and the MetadataExtractor, a dummy field extractor that throws
exceptions is returned for TTL and Timestamp when running against
ES 6+.

Fixes #986