Introduce Gradle buildSrc and prepare the project for a new release manager system.
Moving elasticsearch fixture code to a proper plugin.
Change names of cluster properties.
Remove def types and set types in method signatures.
Dynamically choose whether or not we are building a snapshot at the commandline.
Adding a writePom task that exports a pom to artifacts.
Added a step to the pack task that copies the jars to the artifacts directory.
Assemble depends on writePom now, and also copies the libs to the artifact directory.
Move test jar to build instead of assemble. Change artifacts directory to distributions.