Fixing issues with checking for index, type and document existence.
In 5.5 it seems that the old deprecated way of doing things is now gone.
Added new methods for checking existence of indices, types on indices,
and documents in those types. These calls should be explicit for what
they're checking.

When initializing and checking if the index exists, we need to ensure
that the ES version is discovered now. When testing, make sure the
extended rest client has the version installed. Ignore routing data
on parent child for tests on 1.x.