Adding support for Elasticsearch's new join field to the connector
- Adding the "es.mapping.join" parameter to config.
- Adding extractor class config for join field extractor.
- Added a chained field extractor that checks for routing first, and if
a routing has not been set, will search for the join field's parent id.
- Added some test data for the new join field tests.
- When we hit a join field during mapping parsing, we just create two
meta fields under it and generally treat it as if it were a compound
type. When we hit the long form value it functions as expected, and
when we hit the short form value, we know it's a join field and box it.
- Add skipping ability to FieldExtractors.
- FieldWriter should take an optional header value and output it before
the extracted data if the extractor returns a non "SKIP" response.
- ChainedFieldExtractor can be configured to return either "NOT_FOUND"
or "SKIP" depending on not being able to extract a field from any of its
child extractors.
- Integration Test coverage for join fields