Removing NoRouteToHostException and UnknownHostException
These were added in #505 as fatal network exceptions. While they may
indicate a problem with the network topology, there's still value in
retrying with the next set of nodes until they're exhausted. If the
node is actually unplugged, then it will eventually fail and produce
the same error to the user.

fixes #993