Cleaning up some poor Exception handling in IOUtils.
This throws an unhelpful error when trying to load a resource. First, it
assumes that the resource is being used for the connector's
user-specified-pushdown query. Second, it swallows the root reason for
not being able to open the resource for reading. This is confusing if
you are not loading a file that has anything to do with the pushdown
query, in addition to not having any info about why the connector cannot
retrieve the given resource.

relates #1018