Implement a good chunk of http retry logic.
Things are getting hairy with how much is changing. Should probably
compartmentalize this in a different class at this point for managing
bulk operations.

- Add HttpRetryHandler which instantiates the retry policy and controls
when a message has been retried enough times and how long we should wait
between retries.
- Add method to pass messages along to later handlers letting them know
why a handler decided to not handle a failure.
- Exceptional now lets you return a list of messages from previous
handlers.
- Parameterize the ErrorCollector types used in ErrorHandlers. We want
to be able to extend ErrorCollectors depending on the failure handler
case (one for bulk failures, one for serialization failures).
- Add DelayableErrorCollector which allows you to perform a retry operation
with a wait time between retries.
- Implement DelayableErrorCollector interface in BulkWriteErrorCollector.
- Add passReasons and attemptNumber to BulkWriteFailure object.
- Instantiate the HttpRetryHandler and put it in front of the other
failure handlers.
- Add logging to retry cases.
- Track retry operations and pass messages while evaluating handlers.
- Gut most of the existing HTTP Retry logic. Make it just a pure filter
function instead of a stateful strategy.