Track request attempts per failed document instead of globally.
- Externalize and dumb down the ErrorCollector to just be a value holder
instead of anything that aggregates state. This is because the collector
is only called when a retry occurs, and is not notified when we're
working on a new document.
- Clarify that getPosition is the position of the document in the original
- Lazily create a list of attempts for each document.
- First iteration the list is null since every document will be a 1 value.
- Later iterations, the list will be instantiated with an attempt counter
for each document's position location. (NB: The position changes based on
where it is IN THAT REQUEST. SEE Below)
- Each document that would be retried, check to see if the document changed,
if it did not, increment the attempts counter and add it to the list of attempts.
If it did change, reset the counter to zero and add it to the list.
- After all documents have been processed, the retried documents will be
in the tracking bytes array aligned by position with the attempts in the
attempts-array.
- If further failures are encountered, the attempts-array will be passed
on to the next iteration and used to determine the number of attempts for
documents going forward (as mentioned above).