WIP
Add a basic retry method on error collector to retry current record.
Update the HttpRetryHandler to use this retry method instead of unwrapping the underlying bytes stream.
Add implementations for this new way of retrying data to the bulk write error collector.
Return an input stream for the entry contents when a bulk write fails instead of a byte array. This allows us to avoid spending time copying bytes and provides a top level interface for users to depend on.
Add a blank retry method with a backoff time to the delayed error collector.
BulkError should keep track of the failed document's original position in the list of data entries.
Update all Todos and Fixmes to Fixheres.
Keep track of each document's number of attempts to send to ES. This is important for HTTP Retry logic as there will be a limit on the number of times it will be retried to keep parity with current features.
Add some really sophisticated checks to see if the data actually changed on a retry. If it does not, do not copy bytes in the tracking array.
Change to TrackingBytesArray to allow for getting an entry out of it by index without copying it.