Merged the parsing logic with the error handling logic. Time to start testing.
Keep settings around as a field for if we need to hard reset the tracking bytes array.
Remove final from tracking bytes array byte fields in case we need to hard reset it.
Keep track of if we've potentially grown the tracking bytes array so we can hard reset it.
Track documents sent and skipped at a finer granularity.
We need to keep track of each document's original position in the first tracking bytes array. If a document is modified, it moves to the back of the array, but if it ultimately fails, we need to signal that failure to anything that needs it. Like Storm.
Track BulkErrors as outcomes from ABORT responses, instead of the old way, which was the same as BulkFailures.
Collapse duplicate logic branches to simplify code.
Add a sliceEquals method to ArrayUtils to avoid copying data to check if documents are equal.