Almost done with bulk error handlers.
Move BulkResponse and handler package to the bulk package.
HandlerLoader will now place the failure handler at the end of the list of loaded handlers if it was not explicitly loaded.
Documentation updates for most of the API.
ThrowOnFailure changed to AbortOnFailure, which returns ABORT instead of throwing the original exception.
Fix null check in error collector pass method.
Track attempts explicitly in BulkProcessor flush method.
Label the handler loop in flush method and break to that label from the switch cases.
AbortHandlerException will be treated as an abort result instead of as a regular exception.
Add testing for BulkWriteErrorCollector.
Add extensive testing for the BulkProcessor.