Code cleanup
Add a cause option to abort handler exceptions to be logged in the bulk processor before marking the document as aborted.
Only reset the tracking bytes array if it actually is growing (In cases where the document count threshold is met but not the document space threshold).
Instead of tracking request attempt numbers and original positioning of each entry separately, combine them. They are tracked the same exact way.
Reorder logic branches to make the code read better.
Add much needed comments to flush logic.