Add "Dead Letter Handlers" for bulk write failures (#1095)
New user facing API's have been added in this PR. This includes the error handler interfaces, 
and new settings for specifying error handler class names and error handler properties. Error 
handlers are created at write initialization time and passed handler specific properties. 
Handlers are executed in the order that they were specified in the properties. A handler is 
given the opportunity to handle the error, mark the error as a valid error, or pass the error on to 
the next handler in the chain for processing. This creates a light weight pipeline for handling 
errors that allows users to compose handlers together to increase their job's resiliency to 
failure. When the job concludes, all handlers are closed so that they may clean up any 
resources they may have allocated.