Parse Spark properties in an order agnostic way.
Some spark integrations set both "path" and "resource" in the datasource config, which can cause issues
if they collide. While one of them contains the correct setting, another may hold an invalid setting
which will overwrite the good one based on the order of the map keys. This commit checks for "es.resource",
then "resource", then "path" in that order for values and uses the first found value as the resource to
use.

fixes #1082