Fix timeout not being set for http connections.
Previously, http timeout was being set for sockets and on the
connection manager. This works great if you want the timeout for
receiving data or waiting on the http connection pool, but the
timeout was never used for actually establishing a connection,
which causes slow downs when failing over absent or broken server
hosts.

fixes #1124