When reading metadata from ES, enable tracking scores.
Our scroll requests to Elasticsearch directly specify to use _doc sorting
for performance reasons. Since we are using this sort feature, Elasticsearch
deems it unnecesary to process the scores for documents as they are unused
in the sorting of results. In the case that users are reading metadata, we
should pass the "track_score" parameter into the request so that scores are
returned non-null in the response.

Fixes #1137