First pass at DeserializationErrorHandlers.
Add settings to the ScrollReaderConfig. Include additional info on
Scroll to explicitly denote when a scroll has concluded. This gets
around the scroll being terminated if every document is skipped. If
every document is skipped, continue to advance the scroll until you have
a document to use.

Add DeserializationErrorHandler, the failure class, a loader, and a
collector. Add default FAIL and LOG handlers (copies of the bulk ones).

Load handlers in the ScrollReader. Always copy the content before using
the parser. When reading the response, each time we begin reading a hit
we will wrap the parser in a block aware parser to allow skipping to
the end of the search hit. We note the start and end of a hit in the
main content to slice a byte array out of it for use in the error event.

Add in the retry/handle loop to the readHit method. On retry, wrap the
retry bytes in a new parser and pass it down into the parsing logic.