Unify base error handler implementations (#1163)
Create a base Exceptional class, lots of package moves for handler loaders
and ootb implementations. Interfaces for each handler type are extracted
and used for handler type checking. Delegating handlers allow adapting of
specific interfaces to the general interfaces in a way that still allows for type
checking. New interfaces to make the basic handlers pluggable with behavior
on a per-error-type basis.