Add ability to specify sensitive settings through a keystore (#1174)
ES-Hadoop operates based on the configuration for a given job in Hadoop/Spark. 
In many cases, users need to provide passwords in these configurations. Since 
these settings are kept as plain text, providing them in the job configuration can 
be a potential security hazard. This PR adds support for storing password fields 
in a keystore file, as well as a tool for generating the keystore file.