Use the cluster information in the settings to select ES tokens.
Doesn't magically support multiple clusters, but makes it easier in the
future to determine which cluster a token belongs to. Then a client just
needs to select the right settings for the cluster mentioned in the
token in order to refresh/cancel the token.

Rename the clusername getters and properties to avoid method name
duplication.