Operationalize entire Hadoop cluster fixtures
A cluster config has service configs, which have role configs which
have instance configs. Properties can be specified on all of these
nodes and inherited downward.

Everything that is specific to a service/role is put into a service
descriptor object that defines default values and behaviors as well as
constants for running the service in question.

Cluster formation tasks are now mostly built, but could use some fine
tuning, the likes of which will only become aparent as we use this more.