First step of removing daemon commands.
We need to wrap the process so that we know if it prematurely exits, but
we also need the pid of the process that runs the commands, which
usually runs the exec command. Write a script that execs the start
script after writing its pid to a file, and a script that runs that
wrapper in a child process so it can report on process death.