Allow for gateway roles
A gateway role is the service in question downloaded and extracted
locally, with all configurations needed to launch a job using that
service on the configured cluster.

Gateways do not have running processes, so the tasks are just the
extraction and configuration steps for standing them up.

Tasks will be added later that use those gateways as their launching
point.