Breakout configuration steps per service
Get rid of specifying the data directory. Every service is different.
Add a list of config formats for the config files
Add a step to collect the instance's settings to the service descriptor.
Interrogate the service descriptor for a config format.
HDFS and YARN have some weird dependencies between each other for
selecting config defaults.