Add a build qualifier system property to the build (#1215)
Allows build processes to pick up a qualifier to add into the version
at build time instead of depending on it from a version file. By
default, it will continue to use alpha1 as the qualifier until the ES
build changes the master branch to ship an unqualified snapshot version
during the 7 prerelease period.