Add a breakout to Elasticsearch's cluster conf
We need to have multiple tasks depend on an Elasticsearch cluster
running for the life of the build. In the hadoop fixture, we allow
adding a task or creating a task within the cluster, to be encapsulated
by the cluster start and stop tasks. Do the same thing with the cluster
from Elasticsearch, by adding an extension class for tracking the added
tasks and using a placeholder task to capture the added dependencies
and finalize tasks for use on the provided cluster tasks.