First attempt at Kerberizing Hive.
Beeline requires the Kerberos TGT to be present and makes no attempts at
logging in. To counter this, we monkey patch the beeline script to run
our own SecureBeeline class, which is essentially a shim that performs
a keytab based login, and immediately runs the Beeline class.

Reenable the Hive based tasks, add hive principals to the KDC, and add
the patchBeeline task.

We add the hivePrincipal to the connection string in the task and
accept env parameters for the script.