Parse total hits object from 7.0
Add an equals and hashcode method to BytesArray.
Clean up some stuff in ScrollReaderConfigBuilder.
When parsing hits, check to see if it is either a number or an object.
If it's a number, return it as the total hits.
If it's an object, retrieve the value from it and check that the
relation is "eq", else we can't rely on the count being accurate.