Add AuthenticationMethod and new authentication method setting.
Going forward, users will need to define what kind of authentication
technique they want to use. The issue here is that we cannot assume that
an existing KerberosPrincipal on the current user implies that
Elasticsearch is configured to accept Kerberos authentication.

Default the AuthenticationMethod to be Basic if a username exists in the
settings to allow for backwards compatibility.