Add groovy code for waiting on URL's to be available.
In the normal ES Fixture, we need to wait for the nodes to form a
cluster. To solve that problem we use the ant builder features to use
the wait-for task in ant. This does not work for Hadoop when it secure
mode is enabled, because all the web front ends are configured to use
SSL with self signed certificates and local host names. Instead of using
Ant to build out the URL wait, I've added this Gradle compliant utility
for doing the wait for condition.