Update CommonsHttpTransport to allow for proxy based authentication.
When a user is detected as a proxy user, the proxy user name is captured
and the real user under it is obtained. The real user credentials are
what are used to perform the authentication, and the proxy user's name
is used by the real user to impersonate the proxy user. It is expected
that the real user in this case can run as the given proxy user, and
that a user object exists in ES that the proxy user can be looked up by.

This will mean that for Kerberos and Hive, an administrator will need
to create a proxier role and map hive's principal to it. They will
also need to map each client to a role, as well as add a separate
user for each client that hive can impersonate (since you can't
impersonate kerberos users).