Discover cluster info in bolt and spout constructors.
Will most likely need another pass of determining if this is the best
place to do this. We don't support multiple ES cluster outputs yet, but
Storm is probably the easiest place to do it. That said, our tokens
don't fully support it, and thus, we will have to deal with it.