Docs: Drop inline callouts
Drops the inline callouts from the docs. This is when you write `<1>`
anywhere but the end of a line. Asciidoctor doesn't support them and
we'd very much like to move to Asciidoctor to generate the docs because
it is being actively maintained.

I've taken a shot at removing all of the callouts and everything
compiles but I suspect it doesn't look quite right.