Refine handling of build-tools dependency
This commit refines how the elasticsearch-hadoop build handles the
elasticsearch build-tools dependency. In particular, we change to
correctly depend on a snapshot version if we are building a snapshot
locally, and we also resume using a local repository, so that we can
depend on locally-built versions of build-tools during release and
snapshot builds.