Upgrade to and fix compatibility with Gradle 5.5 (#1333)
This pull request sorts out some issues related to upgrading to Gradle 5.5. We've 
replaced the way we implement "nested" variant builds with a simple Exec task 
instead of the GradleBuild task that has odd behavior due to it running in an 
embedded daemon.

There were some other misc things related to the upgrade and updated/remove 
APIs in Gradle 5.x.