Update the QA project to use test cluster api (#1430)
Also update the project to run fixtures on the java 8 runtime. Java
home can now be set on cluster configs and is used by fixtures.

A cluster can be configured with a dependent ElasticsearchCluster.
This cluster is depended on for any configuration tasks in case its
cluster address is used.

Settings are now stored as objects in cluster configurations instead of
Strings since they may be GStrings with closures inside of them that
must be resolved ONLY at configuration writing time. This is needed to
resolve test cluster http addresses which are only discoverable once
the cluster is started.

In order to make this resolution process simpler, removed most of the
Map<String, String> types in the code and replaced them with the
FileSettings type which can resolve closures as needed when writing the
configuration to a file.

Upgraded Spark QA runtime to 2.3.4 as the older version has been removed
from apache's mirrors.

AbstractClusterTask now extends DefaultTestClustersTask, as it is the
only way to ensure that Elasticsearch is still running when starting the
task.

In order to configure JAVA_HOME consistently, each AbstractClusterTask
now has its environment variables resolved in its super class instead of
with duplicated code in each task.