Share serialized settings and mappings between partition definitions (#1477)
In situations where large pushdown queries, large mappings, and indices with 
many shards are combined, we can inflate the driver program memory very 
quickly as we duplicate the settings and mappings data for each 
ParititionDefinition we create. This change serializes the settings and mapping 
data once and shares the immutable state between all PartitionDefinition 
instances to avoid ballooning memory.