Integrate changes from main line into the build logic.
Now uses a dist project to generate the distribution.
  Set up distElements for variants
Also, javadoc and sources are exposed to upstread projects via producer configurations.
  Set up javadocElements and sourcesElements for variants

Happy accident, it seems that gradle will do all that javadoc and source set up now as an
incubating feature. Not only that, but with the same names we anticipated.

Not all changes have been integrated, but this is a decent checkpoint to commit at.