Update branch to play nice with new javadocElements and sourcesElements configurations.
These new configurations are more akin to distElements than the previous ones, since they contain the JAR artifacts, not the source files that are used to generate them. In this project, we combine the sources for multiple integrations and use the combined sources to generate the artifacts at each level. The existing configurations have been renamed, and we may want to look at removing distElements in the future (if ever) in favor of these new ones.