Cross compile Scala code using Gradle variants (#1521)
This PR introduces Gradle variants into the ES-Hadoop build in order to compile and test Scala code 
across different Scala versions instead of recursively invoking the build from within itself. This PR 
should allow us to upgrade to Scala 2.12 and above as well as newer versions of Spark without 
breaking release artifact contracts outside of a major release.

Also included is the removal of the shared core source directories for the Spark integration. This 
shared source has been refactored into a core project that is cross compiled for the supported 
Spark and Scala versions used in the SQL integration.