Add support for Spark integration on Scala 2.12 (#1589)
This PR adds Scala 2.12 to the list of supported runtimes for the Spark integration. To use 
Scala 2.12 support, the elasticsearch-spark-20_2.12 artifact id should be used when 
downloading dependencies. The default level of support included in the project-wide zip 
archive and the project-wide uber jar will remain Scala 2.11 for the time being. Eventually 
the default version in 8.0 will be updated to a more recent version of Spark/Scala.