Upgrade Hive to 3.x
Downgrade Hadoop version to 3.1.2. We can't test on any higher versions because hive packages an old version of guava that might be loaded first, which breaks much of the Hadoop code.

Force all Hadoop versions to use 3.1.2.

Add a hive-site.xml file to provide the settings to Hive in the test fixture. Expand the test seed in that file to keep the metastore fresh.

Update the HiveEmbeddedServer start up for the new Hive version.

Switch to using the Abstract SerDe class instead of the now removed SerDe interface.

Add code to handle TimestampWritableV2 and DateWritableV2.