Upgrade Hive and Hadoop qa test versions.
Get the Hive qa kerberos tests running again. Hive now requires the use of the schema tool to start. It additionally needs extra setup steps to patch libraries in from Hadoop in order to run. Some changes were needed to the sql scripts as well to account for changes in job planning.