Do not rely on internal build tool logic (#1654)
This PR detangles the hadoop build from the internal classes of elasticsearch build tools.
Mostly it copies over (simplified) plugins and tasks used by hadoop that are not supported
by build tools and not be puslished in the future to be used by external plugins.
In the longterm we probably will regrow the public surface of the elasticsearch build tools but
for now we basically only want to support TestCluster usage and a basic version of the
PluginBuildPlugin to support external projects building elasticsearch plugins.

This is related to https://github.com/elastic/elasticsearch/issues/71593

Updated licenses of build logic sources to apache 2 license