Setting java boot classpath so that scala 2.10 works on gradle 7.2 (#1800)
Gradle 7.2 broke scala 2.10 compilation. The problem seems to be that the java boot classpath jars are
no longer getting into the scala boot classpath. This commit sets it in two ways: (1) it is explicitly set it
in scalaCompileOptions.additionalParameters (to prevent the build from failing) and (2) it is set in options.bootstrapClasspath to prevent warnings from javac that its boot classpath is incompatible with
the source version.