Ignoring missing indices if es.index.read.missing.as.empty is true (#1997)
Fixing the behavior of the es.index.read.missing.as.empty setting if one or more indices exist and
one or more do not. Previously if any index in the list of indices was missing it would return no
results. Now it returns results from indices that do exist and ignores indices that do not exist.
Behavior is unchanged if this property is unset or set to false.